+++ site/dat/docs/Changelog.md
@@ -11,6 +11,7 @@
  - add zipping folders treatment for remote execution
  - add check for resources overlap (data loss danger) for remote execution
  - cleanup config from null values before sending it to the cloud
+ - pull cumulative percentiles from cloud provisioning
 
 ## 1.5.0 <sup>4 may 2016</sup>
  - add [Tsung](Tsung.md) executor